医疗项目，积分使用初稿

+++ src/main/java/com/pepper/project/ch/appointment/mapper/AppointmentMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface AppointmentMapper {
 
-    Appointment selectAppointmentById(Integer id);
+    Appointment selectAppointmentById(Long id);
     List<Appointment> selectAppointmentList(Appointment appointment);
     List<Appointment> selectUserInfoAppointmentList(Appointment appointment);
